feat(sms-alerts): implement useSmsAlertLog hook for fetching SMS alert log entries and enhance SmsAlertsLogCard and SmsAlertsSettingsCard components
